694: Flattening the Google Secrets config
The config was previously defined as an array of objects, which contained a nested array of objects for the secret mappings. This structure does not work well with env var substitution, there is no way to index into the arrays.

As we only require a single store (and FIDC envs only allow a single store), then adding config for a single store with the GOOGLE_SECRET_ prefix

Issue: https://github.com/SecureApiGateway/SecureApiGateway/issues/694