786: Rollback change to oauth-oidc.json Access Token Modification script
Setting the value back to the default that was configured with the empty environment: `Alpha OAuth2 Access Token Modification Script`, we had overridden the config to `OAuth2 Access Token Modification Script`. The realm specific config should be the default.

Note: both scripts are templates that are currently empty. Therefore, this change results in no change in functionality.

https://github.com/SecureApiGateway/SecureApiGateway/issues/786